refactor(agentic-ai): move OpenAI request-customization construction onto backend
OpenAiBackend now owns building its own OpenAiRequestCustomizations
(headers/queryParameters/bodyProperties) via a single
requestCustomizations() method, implemented by OpenAiApiBackend and
OpenAiCustomBackend from their own connection record. Moves
OpenAiRequestCustomizations itself into the model package alongside
OpenAiChatModelConfiguration so the backend can construct it directly,
dropping the family package's from(OpenAiConnection) factory and its
switch over the backend type.
